move InsertOptimizeEngineTest.java
