+++ sharding-core/sharding-core-optimize/src/test/java/org/apache/shardingsphere/core/optimize/AllOptimizerTests.java
@@ -17,6 +17,7 @@
 
 package org.apache.shardingsphere.core.optimize;
 
+import org.apache.shardingsphere.core.optimize.engine.sharding.InsertOptimizeEngineTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
